+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### Unreleased
+### v0.2.0
 
 #### Added